REFACTOR: 컴포넌트 클래스로 리팩토링
