+++ index.html
@@ -75,6 +75,6 @@
       </section>
     </main>
   </body>
-  <script src="data.js"></script>
-  <script src="script.js"></script>
+  <script src="./data.js"></script>
+  <script src="./src/script.js" type="module"></script>
 </html>
