+++ src/components/Component.js
@@ -1,0 +1,11 @@
+export default class Component {
+  constructor(payload) {
+    const { tagName = 'div', className = '', state = {}, props = {} } = payload;
+    this.el = document.createElement(tagName);
+    this.el.className = className;
+    this.state = state;
+    this.props = props;
+    this.render();
+  }
+  render() {}
+}
